Make bot make sense of Discord users changing usernames,
The tool was originally written with the assumption that Discord
usernames can't be changed. For some reason, I never validated
that assumption. Naturally, because this is all part of an April
Fools, people did in fact change usernames and the bot decided
to create separate entries in the users table for them. That means
statistics were split in two.

This rewrite attempts to fix this issue by collecting a lot more
information about general activity in the specific channels for the
Battle Royale. More exactly, any info that would allow a username
to be mapped to a Discord ID. It registers the observations and
then tries to fix any entries where the ID wasn't clear at first "in
post." That is technically not reliable but for our purposes this
should work well enough for now.