Add some manual fixup JSON files.
We found two specific examples of people changing usernames
without the tool being able to pick up on it. Since there are no
other hints currently accessible for us to determine it, I decided to
instead manually insert the information needed for the bot to
make the mapping happen anyways. Long-term, I would love to
get rid of those.